requirements.txt
django == 3.0.3
mysqlclient
#pip install django == 3.0.3 -i https://mirrors.aliyun.com/pypi/simple --trusted-host mirrors.aliyun.com
django-simple-captcha~=0.5.12
redis~=3.5.3
requests~=2.23.0
django-crispy-forms~=1.9.0
xlwt~=1.3.0
XlsxWriter~=1.2.8
future~=0.18.2
httplib2~=0.17.3
six~=1.14.0
DjangoUeditor~=1.8.143